+++ .idea/deploymentTargetSelector.xml
@@ -4,10 +4,10 @@
     <selectionStates>
       <SelectionState runConfigName="app">
         <option name="selectionMode" value="DROPDOWN" />
-        <DropdownSelection timestamp="2025-06-02T10:51:17.458813100Z">
+        <DropdownSelection timestamp="2025-06-02T12:29:03.916138200Z">
           <Target type="DEFAULT_BOOT">
             <handle>
-              <DeviceId pluginId="Default" identifier="serial=192.168.1.19:42628;connection=d2238933" />
+              <DeviceId pluginId="Default" identifier="serial=192.168.1.19:34640;connection=db7ea51c" />
             </handle>
           </Target>
         </DropdownSelection>
